get publisher's game (by id) endpoint, tests, modified publisher service and game repository for this purpose; created Test folder in test solution to have methods for creating entities gathered in one place
