+++ PracticeGamestore/PracticeGamestore.DataAccess/Repositories/Game/GameRepository.cs
@@ -43,7 +43,12 @@ public class GameRepository(GamestoreDbContext context) : IGameRepository
         await UpdateGenreIdsAsync(game.Id, genreIds);
         await UpdatePlatformIdsAsync(game.Id, platformIds);
     }
-    
+
+    public async Task<IEnumerable<Entities.Game>> GetByPublisherIdAsync(Guid id)
+    {
+        return await _gamesNoTracking.Where(g => g.PublisherId == id).ToListAsync();
+    }
+
     private async Task AddPlatformsAsync(Guid gameId, IEnumerable<Guid> platformIds)
     {
         var gamePlatforms = platformIds.Select(id => new GamePlatform
