+++ PracticeGamestore/PracticeGamestore.DataAccess/Repositories/Game/IGameRepository.cs
@@ -7,4 +7,5 @@ public interface IGameRepository
     Task DeleteAsync(Guid id);
     Task<Guid> CreateAsync(Entities.Game game, List<Guid> genreIds, List<Guid> platformIds);
     Task UpdateAsync(Entities.Game game, List<Guid> genreIds, List<Guid> platformIds);
+    Task<IEnumerable<Entities.Game>> GetByPublisherIdAsync(Guid id);
 }
